requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1 